Add unit tests for ast::as_string
Test cases involve each node type as well as a fairly complex
condition string from a complicated falco rule.

Most tests have a unidirectional variant that looks for specific
string representations, as well as a bidirectional variant that
ensures that when the string representation is re-parsed into a second
ast, that the two asts are equal.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>